time오류 수정완료, sell buy 수정완료,  Wallet 테이블 Coins 테이블 형식처럼 만듦

+++ coinAPI/Models/Contract.cs
@@ -5,5 +5,5 @@ public class Contract{
 
     public int id{get;set;}
     public int coinId{get;set;}
-    public int count{get;set;}
+    public UInt16 count{get;set;}
 }}
